+++ gcovr/utils.py
@@ -280,14 +280,13 @@ class AlwaysMatchFilter(Filter):
 
 class DirectoryPrefixFilter(Filter):
     def __init__(self, directory):
-        path = realpath(directory)
-        os_independent_path = get_os_independent_path(path)
+        os_independent_path = get_os_independent_path(directory)
         pattern = re.escape(f"{os_independent_path}/")
         super().__init__(pattern)
 
     def match(self, path: str):
-        realpath = os.path.normpath(path)
-        return super().match(realpath)
+        path = os.path.normpath(path)
+        return super().match(path)
 
 
 def configure_logging() -> None: